parsers/permit-parsing/requirements.txt
# Seattle Owner-Builder Tracker
# Dependencies

# HTTP Client & API
httpx==0.27.0
sodapy==2.2.0

# Browser Automation
playwright==1.41.0
playwright-stealth==1.0.6

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Async Support
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Proxy & Stealth
fake-useragent==1.4.0

# Environment & Config
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# CLI
typer==0.9.0
rich==13.7.0
